+++ src/vtui/mod.rs
@@ -2,5 +2,6 @@ pub mod rich;
 pub mod vtui;
 pub mod session;
 pub mod history;
+pub mod app;
 
 pub use session::enter;
